chore: clean up clippy pedantic + nursery lints workspace-wide
Run cargo fmt and resolve all 224 `clippy::pedantic`/`clippy::nursery`
warnings (now 0). Most are mechanical (inlined format args, map_or /
is_some_and, let-else, write! over format!-append, cast_signed, doc
# Errors/# Panics sections, literal separators, merged match arms).

Where a lint flagged intentional design, suppressed it with a scoped
#[allow] and a justifying comment rather than contorting the code:
- metadata_table: byte-buffer-as-repr(C) pointer casts + bounded offset
  truncations (module-level allow; note the buffer base alignment gap
  documented in the comment is a known, separate concern)
- padding fields, no-op reset(&self) API, opcode-mirroring enum names,
  percentage cast precision, wrapping PC-advance cast, test bindings,
  and a few long test functions

+++ crates/vm/src/vm/metadata_table/arith.rs
@@ -17,6 +17,6 @@ pub const ARITH_METADATA_EXEC_COUNT_OFFSET: usize = offset_of!(ArithMetadata, ex
 
 const _: () = assert!(ARITH_METADATA_STRIDE == 8);
 const _: () = assert!(
-    1 << ARITH_METADATA_STRIDE_SHIFT == ARITH_METADATA_STRIDE as u32,
+    1 << ARITH_METADATA_STRIDE_SHIFT == ARITH_METADATA_STRIDE,
     "ARITH_METADATA_STRIDE_SHIFT must equal log2(ARITH_METADATA_STRIDE)"
 );
